+++ app/models/goggles_db/meeting_event.rb
@@ -4,7 +4,7 @@ module GogglesDb
   #
   # = MeetingEvent model
   #
-  #   - version:  7-0.3.47
+  #   - version:  7-0.5.10
   #   - author:   Steve A.
   #
   class MeetingEvent < ApplicationRecord
@@ -58,37 +58,27 @@ module GogglesDb
     #-- ------------------------------------------------------------------------
     #++
 
-    # Override: include the minimum required 1st-level attributes & associations.
+    # Override: returns the list of single association names (as symbols)
+    # included by <tt>#to_hash</tt> (and, consequently, by <tt>#to_json</tt>).
     #
-    def minimal_attributes
-      super.merge(minimal_associations)
+    def single_associations
+      %i[meeting_session season season_type event_type pool_type stroke_type heat_type]
     end
 
-    # Override: includes *most* of its 1st-level associations into the typical to_json output.
-    def to_json(options = nil)
-      attributes.merge(
-        'meeting_session' => meeting_session.minimal_attributes
-      ).merge(
-        minimal_associations
-      ).merge(
-        'season' => season.minimal_attributes,
-        'season_type' => season_type.minimal_attributes,
-        'meeting_programs' => meeting_programs.map(&:minimal_attributes)
-      ).to_json(options)
+    # Override: returns the list of multiple association names (as symbols)
+    # included by <tt>#to_hash</tt> (and, consequently, by <tt>#to_json</tt>).
+    #
+    def multiple_associations
+      %i[meeting_programs]
     end
 
-    private
-
-    # Returns the "minimum required" hash of associations.
-    def minimal_associations
-      {
-        'display_label' => decorate.display_label,
-        'short_label' => decorate.short_label,
-        'event_type' => event_type.lookup_attributes,
-        'pool_type' => pool_type.lookup_attributes,
-        'stroke_type' => stroke_type.lookup_attributes,
-        'heat_type' => heat_type.lookup_attributes
-      }
+    # Override: include some of the decorated fields in the output.
+    #
+    def minimal_attributes(locale = I18n.locale)
+      super(locale).merge(
+        'display_label' => decorate.display_label(locale),
+        'short_label' => decorate.short_label(locale)
+      )
     end
   end
 end