+++ app/models/goggles_db/meeting_relay_reservation.rb
@@ -4,7 +4,7 @@ module GogglesDb
   #
   # = MeetingRelayReservation model
   #
-  #   - version:  7-0.3.33
+  #   - version:  7-0.5.10
   #   - author:   Steve A.
   #
   # Same properties and methods as MeetingEventReservation, with just a different table name
@@ -44,10 +44,21 @@ module GogglesDb
     #-- ------------------------------------------------------------------------
     #++
 
-    # Override: include the "minimum required" hash of attributes & associations.
+    # Override: include some of the decorated fields in the output.
     #
-    def minimal_attributes
-      super.merge(minimal_associations)
+    def minimal_attributes(locale = I18n.locale)
+      super(locale).merge(
+        'display_label' => decorate.display_label,
+        'short_label' => decorate.short_label,
+        'swimmer_name' => swimmer.complete_name,
+        'swimmer_label' => swimmer.decorate.display_label(locale),
+        'team_name' => team.editable_name,
+        'team_label' => team.decorate.display_label,
+        'event_label' => event_type.label(locale),
+        'category_label' => category_type.decorate.short_label,
+        'category_code' => category_type.code,
+        'gender_code' => gender_type.code
+      )
     end
 
     # Returns a commodity Hash wrapping the essential data that summarizes the Meeting
@@ -60,30 +71,6 @@ module GogglesDb
         'display_label' => meeting.decorate.display_label,
         'short_label' => meeting.decorate.short_label,
         'edition_label' => meeting.edition_label
-      }
-    end
-
-    # Override: includes most relevant data for its 1st-level associations
-    def to_json(options = nil)
-      attributes.merge(
-        'display_label' => decorate.display_label,
-        'short_label' => decorate.short_label,
-        'meeting' => meeting_attributes,
-        'meeting_event' => meeting_event.minimal_attributes,
-        'event_type' => event_type.lookup_attributes,
-        'badge' => badge.minimal_attributes,
-        'team' => team.minimal_attributes,
-        'swimmer' => swimmer.minimal_attributes
-      ).to_json(options)
-    end
-
-    private
-
-    # Returns the "minimum required" hash of associations.
-    def minimal_associations
-      {
-        'meeting_event' => meeting_event.minimal_attributes
-        # (^^ This includes: event_type, stroke_type & heat_type)
       }
     end
   end